requirements.txt
numpy
pandas
scipy
mne
scikit-learn
pyyaml
torch
timm
einops
requests
openneuro-py
zuna
streamlit
plotly
